Fix/client side encryption and query-pipeline (#33142)
### Packages impacted by this PR
@azure/cosmos

### Issues associated with this PR
N/A

### Describe the problem that is addressed by this PR
This PR 
- fixes the copyObject method for encryption and handles 'undefined' and
'bigint' type.
- modifies initializeEncryption to ensure cache get warms up with
encryption only one time.
- fixes the semaphore.leave error in query pipeline 
- adds diagnostics to all the properties and operations
encrypted/decrypted except encrypting query params

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

---------

+++ sdk/cosmosdb/cosmos/test/public/integration/changeFeedIterator.spec.ts
@@ -838,7 +838,9 @@ describe("test changefeed allVersionsAndDeletes mode for entire container", func
     assert.strictEqual(counter, 20, "20 results should be fetched");
   });
   after(async function () {
-    await container.delete();
+    if (container) {
+      await container.delete();
+    }
   });
 });
 
@@ -979,7 +981,9 @@ describe("test changefeed allVersionsAndDeletes mode for a feed range", function
     assert.strictEqual(counter, 5, "5 results should be fetched");
   });
   after(async function () {
-    await container.delete();
+    if (container) {
+      await container.delete();
+    }
   });
 });
 
@@ -1119,7 +1123,9 @@ describe("test changefeed allVersionsAndDeletes mode for a partition key", funct
     assert.strictEqual(counter, 5, "5 results should be fetched");
   });
   after(async function () {
-    await container.delete();
+    if (container) {
+      await container.delete();
+    }
   });
 });
 
@@ -1278,6 +1284,8 @@ describe("test changefeed allVersionsAndDeletes mode for a prefix partition key"
     assert.strictEqual(counter, 10, "10 results should be fetched");
   });
   after(async function () {
-    await container.delete();
+    if (container) {
+      await container.delete();
+    }
   });
 });
